requirements.txt
# ======================================================
# Deep Learning-Based Object Detection Project
# Study Area: Indore District
# Python Dependencies
# ======================================================

# Core Scientific Computing
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scipy==1.11.1

# Geospatial Libraries
GDAL==3.7.1
rasterio==1.3.8
geopandas==0.13.2
shapely==2.0.1
fiona==1.9.4
pyproj==3.6.0

# Deep Learning - TensorFlow/Keras
tensorflow==2.13.0
keras==2.13.1
scikit-learn==1.3.0
scikit-image==0.21.0

# Google Earth Engine
earthengine-api==0.1.367

# Computer Vision
opencv-python==4.8.0.76
Pillow==10.0.0

# Utilities
tqdm==4.66.1
pyyaml==6.0.1
jupyter==1.0.0
ipykernel==6.25.1

# Visualization
plotly==5.16.1
folium==0.14.0

# Optional but Recommended
h5py==3.9.0  # For saving models
albumentations==1.3.1  # Advanced augmentation
segmentation-models==1.0.1  # Pre-built architectures
